IQR {BiocGenerics}	R Documentation

The Interquartile Range

Description

Compute the interquartile range for a vector.

NOTE: This man page is for the IQR S4 generic function defined in the BiocGenerics package. See ?stats::IQR for the default method (defined in the stats package). Bioconductor packages can define specific methods for objects not supported by the default method.

Usage


IQR(x, na.rm = FALSE, type = 7)

Arguments

x, na.rm, type
See ?stats::IQR.

Value

See ?stats::IQR for the value returned by the default method.

Specific methods defined in Bioconductor packages should behave as consistently as possible with the default method.

See Also

stats::IQR for the default IQR method.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


IQR
showMethods("IQR")
selectMethod("IQR", "ANY")  # the default method

[Package BiocGenerics version 0.28.0 Index]