Ontology {BiocGenerics}	R Documentation

Generic Ontology getter

Description

Get the Ontology of an object.

Usage


Ontology(object)

Arguments

object
An object with an Ontology.

See Also

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

Ontology,GOTerms-method in the AnnotationDbi package for an example of a specific Ontology method (defined for GOTerms objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


Ontology
showMethods("Ontology")

library(AnnotationDbi)
showMethods("Ontology")
selectMethod("Ontology", "GOTerms")

[Package BiocGenerics version 0.28.0 Index]