S3-classes-as-S4-classes {BiocGenerics}	R Documentation

S3 classes as S4 classes

Description

Some old-style (aka S3) classes are turned into formally defined (aka S4) classes by the BiocGenerics package. This allows S4 methods defined in Bioconductor packages to use them in their signatures.

Details

S3 classes currently turned into S4 classes:

connection class and subclasses: connection, file, url, gzfile, bzfile, unz, pipe, fifo, sockconn, terminal, textConnection, gzcon. Addtitionally the characterORconnection S4 class is defined as the union of classes character and connection.

others: AsIs, dist

See Also

setOldClass and setClassUnion in the methods package.

[Package BiocGenerics version 0.28.0 Index]