annotation {BiocGenerics}	R Documentation

Accessing annotation information

Description

Get or set the annotation information contained in an object.

Usage

annotation(object, ...)
annotation(object, ...) <- value

Arguments

object
An object containing annotation information.

...
Additional arguments, for use in specific methods.

value
The annotation information to set on object.

See Also

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

annotation,eSet-method in the Biobase package for an example of a specific annotation method (defined for eSet objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples

annotation
showMethods("annotation")

library(Biobase)
showMethods("annotation")
selectMethod("annotation", "eSet")

[Package BiocGenerics version 0.56.0 Index]