aperm {BiocGenerics}	R Documentation

Transposing an array-like object

Description

Transpose an array-like object by permuting its dimensions.

This is a multidimensional generalization of the t() operator used for 2D-transposition.

NOTE: This man page is for the aperm S4 generic function defined in the BiocGenerics package. See ?base::aperm for the default method (defined in the base package). Bioconductor packages can define specific methods for objects not supported by the default method.

Usage

aperm(a, perm, ...)

Arguments

a
An array-like object.

perm, ...
See ?base::aperm for a description of these arguments.

Value

A transposed version of array-like object a, with subscripts permuted as indicated by the perm vector.

See Also

base::aperm for the default aperm method.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

aperm,SVT_SparseArray-method in the SparseArray package for an example of a specific aperm method (defined for SVT_SparseArray objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples

aperm  # note the dispatch on the 'a' arg only
showMethods("aperm")
selectMethod("aperm", "ANY")  # the default method

[Package BiocGenerics version 0.56.0 Index]