as.list {BiocGenerics}	R Documentation

Coerce to a list

Description

Generic function to coerce to a list, if possible.

NOTE: This man page is for the as.list S4 generic function defined in the BiocGenerics package. See ?base::as.list for the default method (defined in the base package). Bioconductor packages can define specific methods for objects not supported by the default method.

Usage

as.list(x, ...)

Arguments

x
The object to coerce.

...
Additional arguments, for use in specific methods.

Value

An ordinary list.

See Also

base::as.list for the default as.list method.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

as.list,List-method in the S4Vectors package for an example of a specific as.list method (defined for List objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples

as.list
showMethods("as.list")
selectMethod("as.list", "ANY")  # the default method

library(S4Vectors)
showMethods("as.list")
## The as.list() method for List objects:
selectMethod("as.list", "List")

[Package BiocGenerics version 0.56.0 Index]