as.vector {BiocGenerics}	R Documentation

Coerce an object into a vector

Description

Attempt to coerce an object into a vector of the specified mode. If the mode is not specified, attempt to coerce to whichever vector mode is considered more appropriate for the class of the supplied object.

NOTE: This man page is for the as.vector S4 generic function defined in the BiocGenerics package. See ?base::as.vector for the default method (defined in the base package). Bioconductor packages can define specific methods for objects not supported by the default method.

Usage


as.vector(x, mode="any")

Arguments

x
The object to coerce.

mode
See ?base::as.vector for a description of this argument.

Value

A vector.

See ?base::as.vector for the value returned by the default method.

Specific methods defined in Bioconductor packages should behave as consistently as possible with the default method.

See Also

base::as.vector for the default as.vector method.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

as.vector,Rle-method and as.vector,AtomicList-method in the S4Vectors and IRanges packages, respectively, for examples of specific as.vector methods (defined for Rle and AtomicList objects, respectively).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


as.vector  # note the dispatch on the 'x' arg only
showMethods("as.vector")
selectMethod("as.vector", "ANY")  # the default method

[Package BiocGenerics version 0.28.0 Index]