boxplot {BiocGenerics}	R Documentation

Box plots

Description

Produce box-and-whisker plot(s) of the given (grouped) values.

NOTE: This man page is for the boxplot S4 generic function defined in the BiocGenerics package. See ?graphics::boxplot for the default method (defined in the graphics package). Bioconductor packages can define specific methods for objects not supported by the default method.

Usage


boxplot(x, ...)

Arguments

x, ...
See ?graphics::boxplot.

Value

See ?graphics::boxplot for the value returned by the default method.

Specific methods defined in Bioconductor packages should behave as consistently as possible with the default method.

See Also

graphics::boxplot for the default boxplot method.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

boxplot,FeatureSet-method in the oligo package for an example of a specific boxplot method (defined for FeatureSet objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


boxplot
showMethods("boxplot")
selectMethod("boxplot", "ANY")  # the default method

[Package BiocGenerics version 0.28.0 Index]