cbind {BiocGenerics}	R Documentation

Combine objects by rows or columns

Description

cbind and rbind take one or more objects and combine them by columns or rows, respectively.

NOTE: This man page is for the cbind and rbind S4 generic functions defined in the BiocGenerics package. See ?base::cbind for the default methods (defined in the base package). Bioconductor packages can define specific methods for objects (typically vector-like or matrix-like) not supported by the default methods.

Usage


cbind(..., deparse.level=1)
rbind(..., deparse.level=1)

Arguments

...
One or more vector-like or matrix-like objects. These can be given as named arguments.

deparse.level
See ?base::cbind for a description of this argument.

Value

See ?base::cbind for the value returned by the default methods.

Specific methods defined in Bioconductor packages will typically return an object of the same class as the input objects.

See Also

base::cbind for the default cbind and rbind methods.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

cbind,DataFrame-method in the S4Vectors package for an example of a specific cbind method (defined for DataFrame objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


cbind  # note the dispatch on the '...' arg only
showMethods("cbind")
selectMethod("cbind", "ANY")  # the default method

rbind  # note the dispatch on the '...' arg only
showMethods("rbind")
selectMethod("rbind", "ANY")  # the default method

[Package BiocGenerics version 0.28.0 Index]