density {BiocGenerics}	R Documentation

Kernel density estimation

Description

The generic function density computes kernel density estimates.

NOTE: This man page is for the density S4 generic function defined in the BiocGenerics package. See ?stats::density for the default method (defined in the stats package). Bioconductor packages can define specific methods for objects not supported by the default method.

Usage


density(x, ...)

Arguments

x, ...
See ?stats::density.

Value

See ?stats::density for the value returned by the default method.

Specific methods defined in Bioconductor packages should behave as consistently as possible with the default method.

See Also

stats::density for the default density method.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

density,flowClust-method in the flowClust package for an example of a specific density method (defined for flowClust objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


density
showMethods("density")
selectMethod("density", "ANY")  # the default method

[Package BiocGenerics version 0.28.0 Index]