dims {BiocGenerics}	R Documentation

Get the dimensions of all elements in a list-like object

Description

Get the dimensions of all elements in a list-like object.

Usage


dims(x)

Arguments

x
List-like object (or environment) containing one or several array-like objects with the same number of dimensions.

See Also

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

dims,eSet-method in the Biobase package for an example of a specific dims method (defined for eSet objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


dims
showMethods("dims")

library(Biobase)
showMethods("dims")
selectMethod("dims", "eSet")

[Package BiocGenerics version 0.28.0 Index]