fileName {BiocGenerics}	R Documentation

Accessing the file name of an object

Description

Get the file name of an object.

Usage


fileName(object, ...)

Arguments

object
An object with a file name.

...
Additional arguments, for use in specific methods.

See Also

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

fileName,MSmap-method in the MSnbase package for an example of a specific fileName method (defined for MSmap objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


fileName
showMethods("fileName")

library(MSnbase)
showMethods("fileName")
selectMethod("fileName", "MSmap")

[Package BiocGenerics version 0.28.0 Index]