format {BiocGenerics}	R Documentation

Format an R object for pretty printing

Description

Turn an R object into a character vector used for pretty printing.

NOTE: This man page is for the format S4 generic function defined in the BiocGenerics package. See ?base::format for the default method (defined in the base package). Bioconductor packages can define specific methods for objects not supported by the default method.

Usage

format(x, ...)

Arguments

x
The object to format.

...
Additional arguments, for use in specific methods.

Value

A character vector that provides a "compact representation" of x. This character vector is typically used by print.data.frame to display the columns of a data.frame object. See ?base::print.data.frame for more information.

See Also

base::format for the default format method.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples

format
showMethods("format")
selectMethod("format", "ANY")  # the default method

[Package BiocGenerics version 0.56.0 Index]