image {BiocGenerics}	R Documentation

Display a color image

Description

Creates a grid of colored or gray-scale rectangles with colors corresponding to the values in z. This can be used to display three-dimensional or spatial data aka images.

NOTE: This man page is for the image S4 generic function defined in the BiocGenerics package. See ?graphics::image for the default method (defined in the graphics package). Bioconductor packages can define specific methods for objects not supported by the default method.

Usage


image(x, ...)

Arguments

x, ...
See ?graphics::image.

Details

See ?graphics::image for the details.

Specific methods defined in Bioconductor packages should behave as consistently as possible with the default method.

See Also

graphics::image for the default image method.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

image,FeatureSet-method in the oligo package for an example of a specific image method (defined for FeatureSet objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


image
showMethods("image")
selectMethod("image", "ANY")  # the default method

[Package BiocGenerics version 0.28.0 Index]