is.unsorted {BiocGenerics}	R Documentation

Test if a vector-like object is not sorted

Description

Test if a vector-like object is not sorted, without the cost of sorting it.

NOTE: This man page is for the is.unsorted S4 generic function defined in the BiocGenerics package. See ?base::is.unsorted for the default method (defined in the base package). Bioconductor packages can define specific methods for objects (typically vector-like) not supported by the default method.

Usage


is.unsorted(x, na.rm=FALSE, strictly=FALSE, ...)

Arguments

x
A vector-like object.

na.rm, strictly
See ?base::is.unsorted for a description of these arguments.

...
Additional arguments, for use in specific methods.

Note that base::is.unsorted (the default method) only takes the x, na.rm, and strictly arguments.

Value

See ?base::is.unsorted for the value returned by the default method.

Specific methods defined in Bioconductor packages should behave as consistently as possible with the default method.

Note

TO DEVELOPERS:

The is.unsorted method for specific vector-like objects should adhere to the same underlying order used by the order, sort, and rank methods for the same objects.

See Also

base::is.unsorted for the default is.unsorted method.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

is.unsorted,GenomicRanges-method in the GenomicRanges package for an example of a specific is.unsorted method (defined for GenomicRanges objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


is.unsorted  # note the dispatch on the 'x' arg only
showMethods("is.unsorted")
selectMethod("is.unsorted", "ANY")  # the default method

[Package BiocGenerics version 0.28.0 Index]