longForm {BiocGenerics}	R Documentation

Turn object into long form

Description

A generic function that returns the long form of an object.

Usage

longForm(object, ...)

Arguments

object
A data object.

...
Additional arguments, for use in specific methods.

Value

The long form version of the original object. This is typically a data-frame-like object.

See Also

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

longForm,MultiAssayExperiment-method in the MultiAssayExperiment package for an example of a specific longForm method (defined for MultiAssayExperiment objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples

longForm
showMethods("longForm")

library(MultiAssayExperiment)
showMethods("longForm")
selectMethod("longForm", "MultiAssayExperiment")

[Package BiocGenerics version 0.56.0 Index]