normalize {BiocGenerics}	R Documentation

Normalize an object

Description

A generic function which normalizes an object containing microarray data or other data. Normalization is intended to remove from the intensity measures any systematic trends which arise from the microarray technology rather than from differences between the probes or between the target RNA samples hybridized to the arrays.

Usage


normalize(object, ...)

Arguments

object
A data object, typically containing microarray data.

...
Additional arguments, for use in specific methods.

Value

An object containing the normalized data.

See Also

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

normalize,AffyBatch-method in the affy package and normalize,FeatureSet-method in the oligo package for examples of specific normalize methods (defined for AffyBatch and FeatureSet objects, respectively).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


normalize
showMethods("normalize")

library(affy)
showMethods("normalize")
selectMethod("normalize", "AffyBatch")

[Package BiocGenerics version 0.28.0 Index]