plotMA {BiocGenerics}	R Documentation

MA-plot: plot differences versus averages for high-throughput data

Description

A generic function which produces an MA-plot for an object containing microarray, RNA-Seq or other data.

Usage


plotMA(object, ...)

Arguments

object
A data object, typically containing count values from an RNA-Seq experiment or microarray intensity values.

...
Additional arguments, for use in specific methods.

Value

Undefined. The function exists for its side effect, producing a plot.

See Also

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

plotMA in the limma package for a function with the same name that is not dispatched through this generic function.

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


showMethods("plotMA")

suppressWarnings(
  if(require("DESeq2"))
    example("plotMA", package="DESeq2", local=TRUE)
)

[Package BiocGenerics version 0.28.0 Index]