rank {BiocGenerics}	R Documentation

Ranks the values in a vector-like object

Description

Returns the ranks of the values in a vector-like object. Ties (i.e., equal values) and missing values can be handled in several ways.

NOTE: This man page is for the rank S4 generic function defined in the BiocGenerics package. See ?base::rank for the default method (defined in the base package). Bioconductor packages can define specific methods for objects not supported by the default method.

Usage


rank(x, na.last=TRUE,
     ties.method=c("average", "first", "last", "random", "max", "min"),
     ...)

Arguments

x
A vector-like object.

na.last, ties.method
See ?base::rank for a description of these arguments.

...
Additional arguments, for use in specific methods.

Note that base::rank (the default method) only takes the x, na.last, and ties.method arguments.

Value

See ?base::rank for the value returned by the default method.

Specific methods defined in Bioconductor packages should behave as consistently as possible with the default method.

Note

TO DEVELOPERS:

See note in ?BiocGenerics::order about "stable" order.

order, sort, and rank methods for specific vector-like objects should adhere to the same underlying order that should be conceptually defined as a binary relation on the set of all possible vector values. For completeness, this binary relation should also be incarnated by a <= method.

See Also

base::rank for the default rank method.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

rank,Vector-method in the S4Vectors package for an example of a specific rank method (defined for Vector objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


rank  # note the dispatch on the 'x' arg only
showMethods("rank")
selectMethod("rank", "ANY")  # the default method

[Package BiocGenerics version 0.28.0 Index]