rep {BiocGenerics}	R Documentation

Replicate elements of a vector-like object

Description

rep.int replicates the elements in x.

NOTE: This man page is for the rep.int S4 generic function defined in the BiocGenerics package. See ?base::rep.int for the default method (defined in the base package). Bioconductor packages can define specific methods for objects (typically vector-like) not supported by the default method.

Usage


rep.int(x, times)

Arguments

x
The object to replicate (typically vector-like).

times
See ?base::rep.int for a description of this argument.

Value

See ?base::rep.int for the value returned by the default method.

Specific methods defined in Bioconductor packages will typically return an object of the same class as the input object.

See Also

base::rep.int for the default rep.int, intersect, and setdiff methods.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

rep.int,Rle-method in the S4Vectors package for an example of a specific rep.int method (defined for Rle objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


rep.int
showMethods("rep.int")
selectMethod("rep.int", "ANY")  # the default method

[Package BiocGenerics version 0.28.0 Index]