residuals {BiocGenerics}	R Documentation

Extract model residuals

Description

residuals is a generic function which extracts model residuals from objects returned by modeling functions.

NOTE: This man page is for the residuals S4 generic function defined in the BiocGenerics package. See ?stats::residuals for the default method (defined in the stats package). Bioconductor packages can define specific methods for objects not supported by the default method.

Usage


residuals(object, ...)

Arguments

object, ...
See ?stats::residuals.

Value

Residuals extracted from the object object.

See Also

stats::residuals for the default residuals method.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

residuals,PLMset-method in the affyPLM package for an example of a specific residuals method (defined for PLMset objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


residuals
showMethods("residuals")
selectMethod("residuals", "ANY")  # the default method

[Package BiocGenerics version 0.28.0 Index]