saveRDS {BiocGenerics}	R Documentation

The saveRDS() S4 generic and default method

Description

Generic function to write a single R object to a file.

NOTE: This man page is for the saveRDS S4 generic function and default method defined in the BiocGenerics package. See ?base::saveRDS for the corresponding function defined in base R.

Usage

saveRDS(object, file="", ascii=FALSE, version=NULL,
        compress=TRUE, refhook=NULL)

Arguments

object, file, ascii, version, compress, refhook
See ?base::saveRDS for a description of these arguments.

Details

The default saveRDS method defined in this package is a thin wrapper around base::saveRDS that issues a warning if the object to serialize contains out-of-memory data. See ?containsOutOfMemoryData for more information.

Bioconductor packages can override this default method with more specialized methods.

Value

An invisible NULL.

See Also

base::saveRDS in the base package for the default saveRDS method.

containsOutOfMemoryData for determining whether an object contains out-of-memory data or not.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

saveRDS,SummarizedExperiment-method in the SummarizedExperiment package for an example of a specific saveRDS method (defined for SummarizedExperiment objects and derivatives).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples

saveRDS  # note the dispatch on the 'object' arg only
showMethods("saveRDS")
selectMethod("saveRDS", "ANY")  # the default method

[Package BiocGenerics version 0.56.0 Index]