score {BiocGenerics}	R Documentation

Score accessor

Description

Get or set the score value contained in an object.

Usage


score(x, ...)
score(x, ...) <- value

Arguments

x
An object to get or set the score value of.

...
Additional arguments, for use in specific methods.

value
The score value to set on x.

See Also

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

score,GenomicRanges-method in the GenomicRanges package for an example of a specific score method (defined for GenomicRanges objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


score
showMethods("score")

`score<-`
showMethods("score<-")

library(GenomicRanges)

showMethods("score")
selectMethod("score", "GenomicRanges")

showMethods("score<-")
selectMethod("score<-", "GenomicRanges")

[Package BiocGenerics version 0.28.0 Index]