strand {BiocGenerics}	R Documentation

Accessing strand information

Description

Get or set the strand information contained in an object.

Usage


strand(x, ...)
strand(x, ...) <- value

unstrand(x)

invertStrand(x)
## S4 method for signature 'ANY'
invertStrand(x)

Arguments

x
An object containing strand information.

...
Additional arguments, for use in specific methods.

value
The strand information to set on x.

Details

All the strand methods defined in the GenomicRanges package use the same set of 3 values (called the "standard strand levels") to specify the strand of a genomic location: +, -, and *. * is used when the exact strand of the location is unknown, or irrelevant, or when the "feature" at that location belongs to both strands.

Note that unstrand is not a generic function, just a convenience wrapper to the generic strand() setter (strand<-) that does:

  strand(x) <- "*"
  x

The default method for invertStrand does:

  strand(x) <- invertStrand(strand(x))
  x

Value

If x is a vector-like object, strand(x) will typically return a vector-like object parallel to x, that is, an object of the same length as x where the i-th element describes the strand of the i-th element in x.

unstrand(x) and invertStrand(x) return a copy of x with the strand set to "*" for unstrand or inverted for invertStrand (i.e. "+" and "-" switched, and "*" untouched).

See Also

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

strand,GRanges-method in the GenomicRanges package for an example of a specific strand method (defined for GRanges objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


strand
showMethods("strand")

`strand<-`
showMethods("strand<-")

unstrand

invertStrand
showMethods("invertStrand")
selectMethod("invertStrand", "ANY")  # the default method

library(GenomicRanges)

showMethods("strand")
selectMethod("strand", "missing")
strand()

showMethods("strand<-")

[Package BiocGenerics version 0.28.0 Index]