table {BiocGenerics}	R Documentation

Cross tabulation and table creation

Description

table uses the cross-classifying factors to build a contingency table of the counts at each combination of factor levels.

NOTE: This man page is for the table S4 generic function defined in the BiocGenerics package. See ?base::table for the default method (defined in the base package). Bioconductor packages can define specific methods for objects not supported by the default method.

Usage


table(...)

Arguments

...
One or more objects which can be interpreted as factors (including character strings), or a list (or data frame) whose components can be so interpreted.

Value

See ?base::table for the value returned by the default method.

Specific methods defined in Bioconductor packages should also return the type of object returned by the default method.

See Also

base::table for the default table method.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

table,Rle-method in the S4Vectors package for an example of a specific table method (defined for Rle objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


table
showMethods("table")
selectMethod("table", "ANY")  # the default method

[Package BiocGenerics version 0.28.0 Index]