toTable {BiocGenerics}	R Documentation

An alternative to as.data.frame()

Description

toTable() is an S4 generic function provided as an alternative to as.data.frame().

Usage

toTable(x, ...)

Arguments

x
The object to turn into a data frame.

...
Additional arguments, for use in specific methods.

Value

A data frame.

See Also

The as.data.frame S4 generic defined in the BiocGenerics package.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

toTable,Bimap-method in the AnnotationDbi package for an example of a specific toTable method (defined for Bimap objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples

toTable
showMethods("toTable")

library(AnnotationDbi)
showMethods("toTable")
selectMethod("toTable", "Bimap")

[Package BiocGenerics version 0.36.1 Index]