var {BiocGenerics}	R Documentation

Variance and Standard Deviation

Description

var and sd compute the variance and standard deviation of a vector x.

NOTE: This man page is for the var and sd, S4 generic functions defined in the BiocGenerics package. See ?stats::var and ?stats::sd for the default methods (defined in the stats package). Bioconductor packages can define specific methods for objects (typically array-like) not supported by the default method.

Usage


     var(x, y = NULL, na.rm = FALSE, use)
     sd(x, na.rm = FALSE)

Arguments

x
a vector-like object

y
a vector-like object, or NULL

na.rm, use
see var

Value

See ?stats::var and ?stats::sd for the value returned by the default methods.

Specific methods defined in Bioconductor packages will typically return an object of the same class as the input object.

See Also

stats::var and stats::sd for the default methods.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


var
showMethods("var")
selectMethod("var", "ANY")  # the default method

[Package BiocGenerics version 0.28.0 Index]