weights {BiocGenerics}	R Documentation

Extract model weights

Description

weights is a generic function which extracts fitting weights from objects returned by modeling functions.

NOTE: This man page is for the weights S4 generic function defined in the BiocGenerics package. See ?stats::weights for the default method (defined in the stats package). Bioconductor packages can define specific methods for objects not supported by the default method.

Usage


weights(object, ...)

Arguments

object, ...
See ?stats::weights.

Value

Weights extracted from the object object.

See ?stats::weights for the value returned by the default method.

Specific methods defined in Bioconductor packages should behave as consistently as possible with the default method.

See Also

stats::weights for the default weights method.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

weights,PLMset-method in the affyPLM package for an example of a specific weights method (defined for PLMset objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples


weights
showMethods("weights")
selectMethod("weights", "ANY")  # the default method

[Package BiocGenerics version 0.28.0 Index]