as.data.frame {BiocGenerics}	R Documentation

Coerce to a data frame

Description

Generic function to coerce to a data frame, if possible.

NOTE: This man page is for the as.data.frame S4 generic function defined in the BiocGenerics package. See ?base::as.data.frame for the default method (defined in the base package). Bioconductor packages can define specific methods for objects not supported by the default method.

Usage

as.data.frame(x, row.names=NULL, optional=FALSE, ...)

Arguments

x
The object to coerce.

row.names, optional, ...
See ?base::as.data.frame for a description of these arguments.

Value

An ordinary data frame.

See ?base::as.data.frame for the value returned by the default method.

Specific methods defined in Bioconductor packages should behave as consistently as possible with the default method.

See Also

base::as.data.frame for the default as.data.frame method.

toTable for an alternative to as.data.frame.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

as.data.frame,DataFrame-method in the S4Vectors package, and as.data.frame,IntegerRanges-method in the IRanges package, for examples of specific as.data.frame methods (defined for DataFrame and IntegerRanges objects, respectively).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples

as.data.frame  # note the dispatch on the 'x' arg only
showMethods("as.data.frame")
selectMethod("as.data.frame", "ANY")  # the default method

[Package BiocGenerics version 0.56.0 Index]