dge {BiocGenerics}	R Documentation

Accessors and generic functions used in the context of count datasets

Description

These generic functions provide basic interfaces to operations on and data access to count datasets.

Usage

counts(object, ...)
counts(object, ...) <- value
design(object, ...)
design(object, ...) <- value
dispTable(object, ...)
dispTable(object, ...)  <- value
sizeFactors(object, ...)
sizeFactors(object, ...) <- value
conditions(object, ...)
conditions(object, ...) <- value
estimateSizeFactors(object, ...)
estimateDispersions(object, ...)
plotDispEsts(object, ...)

Arguments

object
Object of class for which methods are defined, e.g., CountDataSet, DESeqSummarizedExperiment or ExonCountSet.

value
Value to be assigned to corresponding components of object; supported types depend on method implementation.

...
Further arguments, perhaps used by metohds

Details

For the details, please consult the manual pages of the methods in the DESeq, DESeq2, and DEXSeq packages and the package vignettes.

Author(s)

W. Huber, S. Anders

See Also

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

[Package BiocGenerics version 0.56.0 Index]