do.call {BiocGenerics}	R Documentation

Execute a function call

Description

do.call constructs and executes a function call from a name or a function and a list of arguments to be passed to it.

NOTE: This man page is for the do.call S4 generic function defined in the BiocGenerics package. See ?base::do.call for the default method (defined in the base package). Bioconductor packages can define specific methods for objects not supported by the default method.

Usage

do.call(what, args, quote=FALSE, envir=parent.frame())

Arguments

what
The default method expects either a function or a non-empty character string naming the function to be called. See ?base::do.call for the details.

Specific methods can support other objects. Please refer to the documentation of a particular method for the details.

args
The default method expects a list of arguments to the function call (the names attribute of args gives the argument names). See ?base::do.call for the details.

Specific methods can support other objects. Please refer to the documentation of a particular method for the details.

quote, envir
See ?base::do.call for a description of these arguments.

Value

The result of the (evaluated) function call.

Specific methods defined in Bioconductor packages should behave as consistently as possible with the default method.

See Also

base::do.call for the default do.call method.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples

do.call  # note the dispatch on the 'what' and 'args' args only
showMethods("do.call")
selectMethod("do.call", c("ANY", "ANY"))  # the default method

[Package BiocGenerics version 0.56.0 Index]