evalq {BiocGenerics}	R Documentation

Evaluate an (unevaluated) expression

Description

evalq evaluates an R expression (the quoted form of its first argument) in a specified environment.

NOTE: This man page is for the evalq wrapper defined in the BiocGenerics package. See ?base::evalq for the function defined in the base package. This wrapper correctly delegates to the eval generic, rather than base::eval.

Usage

evalq(expr, envir=parent.frame(),
           enclos=if (is.list(envir) || is.pairlist(envir))
                  parent.frame() else baseenv())

Arguments

expr
Quoted to form the expression that is evaluated.

envir
The environment in which expr is to be evaluated. May be any object supported by methods on the eval generic.

enclos
See ?base::evalq for a description of this argument.

Value

See ?base::evalq.

See Also

base::evalq for the base evalq function.

Examples

evalq  # note just a copy of the original evalq

[Package BiocGenerics version 0.56.0 Index]