paste {BiocGenerics}	R Documentation

Concatenate strings

Description

paste concatenates vectors of strings or vector-like objects containing strings.

NOTE: This man page is for the paste S4 generic function defined in the BiocGenerics package. See ?base::paste for the default method (defined in the base package). Bioconductor packages can define specific methods for objects (typically vector-like objects containing strings) not supported by the default method.

Usage

paste(..., sep=" ", collapse=NULL, recycle0=FALSE)

Arguments

...
One or more vector-like objects containing strings.

sep, collapse, recycle0
See ?base::paste for a description of these arguments.

Value

See ?base::paste for the value returned by the default method.

Specific methods defined in Bioconductor packages will typically return an object of the same class as the input objects.

See Also

base::paste for the default paste method.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

paste,Rle-method in the S4Vectors package for an example of a specific paste method (defined for Rle objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples

paste
showMethods("paste")
selectMethod("paste", "ANY")  # the default method

[Package BiocGenerics version 0.56.0 Index]