t {BiocGenerics}	R Documentation

Matrix Transpose

Description

Given a rectangular object x, t returns the transpose of x.

NOTE: This man page is for the t S4 generic function defined in the BiocGenerics package. See ?base::t for the default method (defined in the base package). Bioconductor packages can define specific methods for objects (typically array-like) not supported by the default method.

Usage

t(x)

Arguments

x
A matrix-like or other rectangular object.

Value

See ?base::t for the value returned by the default method.

Specific methods defined in Bioconductor packages will typically return an object of the same class as the input object.

See Also

base::t for the default t method.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

t,Hits-method in the S4Vectors package for an example of a specific t method (defined for Hits objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples

t
showMethods("t")
selectMethod("t", "ANY")  # the default method

[Package BiocGenerics version 0.56.0 Index]