tapply {BiocGenerics}	R Documentation

Apply a function over a ragged array

Description

tapply applies a function to each cell of a ragged array, that is to each (non-empty) group of values given by a unique combination of the levels of certain factors.

NOTE: This man page is for the tapply S4 generic function defined in the BiocGenerics package. See ?base::tapply for the default method (defined in the base package). Bioconductor packages can define specific methods for objects (typically list-like or vector-like) not supported by the default method.

Usage

tapply(X, INDEX, FUN=NULL, ..., default=NA, simplify=TRUE)

Arguments

X
The default method expects an atomic object, typically a vector. See ?base::tapply for the details.

Specific methods can support other objects (typically list-like or vector-like). Please refer to the documentation of a particular method for the details.

INDEX
The default method expects a list of one or more factors, each of same length as X. See ?base::tapply for the details.

Specific methods can support other objects (typically list-like). Please refer to the documentation of a particular method for the details.

FUN, ..., default, simplify
See ?base::tapply for a description of these arguments.

Value

See ?base::tapply for the value returned by the default method.

Specific methods defined in Bioconductor packages should behave as consistently as possible with the default method.

See Also

base::tapply for the default tapply method.

showMethods for displaying a summary of the methods defined for a given generic function.

selectMethod for getting the definition of a specific method.

tapply,Vector,ANY-method in the IRanges package for an example of a specific tapply method (defined for Vector objects).

BiocGenerics for a summary of all the generics defined in the BiocGenerics package.

Examples

tapply  # note the dispatch on the 'X' and 'INDEX' args only
showMethods("tapply")
selectMethod("tapply", c("ANY", "ANY"))  # the default method

[Package BiocGenerics version 0.56.0 Index]